Small error fixes + checks to speed up reruns

+++ s5_r2/run.sh
@@ -39,6 +39,10 @@ g2p_dir=data/local/g2p${dict_suffix}
 lm_dir=data/local/lm${dict_suffix}
 arpa_lm=${lm_dir}/4gram-mincount/lm_pr10.0.gz
 
+[ ! -L "steps" ] && ln -s ../../wsj/s5/steps
+[ ! -L "utils" ] && ln -s ../../wsj/s5/utils
+[ ! -L "rnnlm" ] && ln -s ../../../scripts/rnnlm/
+
 . utils/parse_options.sh
 
 if [ -f cmd.sh ]; then
@@ -56,10 +60,6 @@ else
     echo "E.g. wget https://www-i6.informatik.rwth-aachen.de/web/Software/g2p-r1668-r3.tar.gz"
     exit
 fi
-
-[ ! -L "steps" ] && ln -s ../../wsj/s5/steps
-[ ! -L "utils" ] && ln -s ../../wsj/s5/utils
-[ ! -L "rnnlm" ] && ln -s ../../../scripts/rnnlm/
 
 # mfccdir should be some place with a largish disk where you
 # want to store MFCC features.
@@ -122,10 +122,13 @@ if [ $stage -le 1 ]; then
     if [ ! -d data/wav/m_ailabs/ ]
     then
       mkdir -p data/wav/m_ailabs/
-      wget --directory-prefix=data/wav/ http://speech.tools/kaldi_tuda_de/m-ailabs.bayern.de_DE.tgz
+      wget --directory-prefix=data/wav/m_ailabs/ http://speech.tools/kaldi_tuda_de/m-ailabs.bayern.de_DE.tgz
       cd data/wav/m_ailabs/
       tar xvfz m-ailabs.bayern.de_DE.tgz
       cd ../../../
+    fi
+    if [ ! -d data/m_ailabs_train ]
+    then
       # make data directory data/m_ailabs_train 
       python3 local/prepare_m-ailabs_data.py
     fi
@@ -139,6 +142,9 @@ RAWDATA=data/wav/german-speechdata-package-v2
 FILTERBYNAME="*.xml"
 
 if [ $stage -le 2 ]; then
+  # remove files, which would later produce errors
+  python3 local/remove_files_to_skip.py data/wav/german-speechdata-package-v2/train/
+
   find $RAWDATA/*/$FILTERBYNAME -type f > data/waveIDs.txt
 
   # prepares directories in Kaldi format for the TUDA speech corpus
@@ -227,33 +233,38 @@ if [ $stage -le 5 ]; then
       echo "G2P model file $final_g2p_model already exists, not recreating it."
   fi
 
-  echo "Now finding OOV in train"
+  if [ ! -f ${dict_dir}/oov_lexiconp.txt ]
+  then
+    echo "Now finding OOV in train"
 
-  cp data/tuda_train/text ${g2p_dir}/complete_text
-  
-  if [ "$add_swc_data" = true ] ; then
-    cat data/swc_train/text >> ${g2p_dir}/complete_text
-  fi  
+    cp data/tuda_train/text ${g2p_dir}/complete_text
 
-  if [ "$add_mailabs_data" = true ] ; then
-    cat data/m_ailabs_train/text >> ${g2p_dir}/complete_text
+    if [ "$add_swc_data" = true ] ; then
+      cat data/swc_train/text >> ${g2p_dir}/complete_text
+    fi
+
+    if [ "$add_mailabs_data" = true ] ; then
+      cat data/m_ailabs_train/text >> ${g2p_dir}/complete_text
+    fi
+
+    if [ "$add_extra_words" = true ] ; then
+      # source extra words from $extra_words_file (e.g. local/extra_words.txt) and prefix them with bogus ids, so that we can just add them to the transcriptions (${g2p_dir}/complete_text)
+      gawk "{ printf(\"extra-word-%i %s\n\",NR,\$1) }" $extra_words_file | cat ${g2p_dir}/complete_text - > ${g2p_dir}/complete_text_new
+      mv ${g2p_dir}/complete_text_new ${g2p_dir}/complete_text
+    fi
+
+    python3 local/find_oov.py -c ${g2p_dir}/complete_text -w ${g2p_dir}/lexicon_wordlist.txt -o ${g2p_dir}/oov.txt
+
+    echo "Now using G2P to predict OOV"
+    $sequitur_g2p -e utf8 --model $final_g2p_model --apply ${g2p_dir}/oov.txt > ${dict_dir}/oov_lexicon.txt
+    cat ${dict_dir}/oov_lexicon.txt | gawk '{$1=$1" 1.0"; print }' > ${dict_dir}/_oov_lexiconp.txt
+    # remove entries that don't have atleast 3 columns - some phonemizations are broken
+    gawk 'NF>=3' ${dict_dir}/_oov_lexiconp.txt > ${dict_dir}/oov_lexiconp.txt
+    #${dict_dir}/oov_lexicon.txt
+    echo "Done!"
+  else
+    echo "OOV already predicted and exists, not recreating."
   fi
-
-  if [ "$add_extra_words" = true ] ; then
-    # source extra words from $extra_words_file (e.g. local/extra_words.txt) and prefix them with bogus ids, so that we can just add them to the transcriptions (${g2p_dir}/complete_text)
-    awk "{ printf(\"extra-word-%i %s\n\",NR,\$1) }" $extra_words_file | cat ${g2p_dir}/complete_text - > ${g2p_dir}/complete_text_new
-    mv ${g2p_dir}/complete_text_new ${g2p_dir}/complete_text
-  fi
-
-  python3 local/find_oov.py -c ${g2p_dir}/complete_text -w ${g2p_dir}/lexicon_wordlist.txt -o ${g2p_dir}/oov.txt
-  
-  echo "Now using G2P to predict OOV"
-  $sequitur_g2p -e utf8 --model $final_g2p_model --apply ${g2p_dir}/oov.txt > ${dict_dir}/oov_lexicon.txt
-  cat ${dict_dir}/oov_lexicon.txt | awk '{$1=$1" 1.0"; print }' > ${dict_dir}/_oov_lexiconp.txt
-  # remove entries that don't have atleast 3 columns - some phonemizations are broken
-  awk 'NF>=3' ${dict_dir}/_oov_lexiconp.txt > ${dict_dir}/oov_lexiconp.txt
-  #${dict_dir}/oov_lexicon.txt
-  echo "Done!"
 fi
 
 # Now start preprocessing with KALDI scripts
@@ -324,7 +335,7 @@ if [ "$add_swc_data" = true ] ; then
       rm data/swc_train/spk2utt
       
       cat data/swc_train/segments | sort > data/swc_train/segments_sorted
-      cat data/swc_train/text | sort | awk 'NF>=2' > data/swc_train/text_sorted
+      cat data/swc_train/text | sort | gawk 'NF>=2' > data/swc_train/text_sorted
       cat data/swc_train/utt2spk | sort > data/swc_train/utt2spk_sorted
       cat data/swc_train/wav.scp | sort > data/swc_train/wav.scp_sorted
 
@@ -367,7 +378,7 @@ if [ "$add_mailabs_data" = true ] ; then
     mv data/train data/train_without_mailabs 
     echo "Now computing MFCC features for m_ailabs_train"
     # Now make MFCC features.
-    x=data/m_ailabs_train
+    x=m_ailabs_train
     utils/fix_data_dir.sh data/$x # some files fail to get mfcc for many reasons
     steps/make_mfcc.sh --cmd "$train_cmd" --nj $nJobs data/$x exp/make_mfcc/$x $mfccdir
     utils/fix_data_dir.sh data/$x # some files fail to get mfcc for many reasons
